refactor: move ClientId to dedicated mod
