+++ crypto/src/mls/client/mod.rs
@@ -14,6 +14,8 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see http://www.gnu.org/licenses/.
 
+pub(crate) mod id;
+
 use openmls::{
     credentials::CredentialBundle,
     extensions::Extension,
@@ -31,71 +33,11 @@ use core_crypto_keystore::{
 };
 use mls_crypto_provider::MlsCryptoProvider;
 
-use crate::{mls::CertificateBundle, mls::MlsCiphersuite, CryptoError, CryptoResult, MlsError};
+use crate::{mls::CertificateBundle, mls::MlsCiphersuite, prelude::ClientId, CryptoError, CryptoResult, MlsError};
 
 pub(crate) const INITIAL_KEYING_MATERIAL_COUNT: usize = 100;
 
 const KEYPACKAGE_DEFAULT_LIFETIME: std::time::Duration = std::time::Duration::from_secs(60 * 60 * 24 * 90); // 3 months
-
-/// A unique identifier for clients. A client is an identifier for each App a user is using, such as desktop,
-/// mobile, etc. Users can have multiple clients.
-/// More information [here](https://messaginglayersecurity.rocks/mls-architecture/draft-ietf-mls-architecture.html#name-group-members-and-clients)
-#[derive(Debug, Clone, PartialEq, Eq, Hash, derive_more::Deref)]
-pub struct ClientId(pub(crate) Vec<u8>);
-
-impl From<&[u8]> for ClientId {
-    fn from(value: &[u8]) -> Self {
-        Self(value.into())
-    }
-}
-
-impl From<Vec<u8>> for ClientId {
-    fn from(value: Vec<u8>) -> Self {
-        Self(value)
-    }
-}
-
-impl From<Box<[u8]>> for ClientId {
-    fn from(value: Box<[u8]>) -> Self {
-        Self(value.into())
-    }
-}
-
-impl From<ClientId> for Box<[u8]> {
-    fn from(value: ClientId) -> Self {
-        value.0.into_boxed_slice()
-    }
-}
-
-#[cfg(test)]
-impl From<&str> for ClientId {
-    fn from(value: &str) -> Self {
-        Self(value.as_bytes().into())
-    }
-}
-
-#[allow(clippy::from_over_into)]
-impl Into<Vec<u8>> for ClientId {
-    fn into(self) -> Vec<u8> {
-        self.0
-    }
-}
-
-impl std::fmt::Display for ClientId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", hex::encode(self.0.as_slice()))
-    }
-}
-
-impl std::str::FromStr for ClientId {
-    type Err = CryptoError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(Self(
-            hex::decode(s).map_or_else(|_| s.as_bytes().to_vec(), std::convert::identity),
-        ))
-    }
-}
 
 /// Represents a MLS client which in our case is the equivalent of a device.
 /// It can be the Android, iOS, web or desktop application which the authenticated user is using.
